configure conditional compile vars with configuration file. (#187)
