+++ src/components/abciapp/src/abci/staking/whoami.rs
@@ -5,7 +5,7 @@
 //!
 
 use {
-    crate::abci::config::global_cfg::CFG, lazy_static::lazy_static,
+    config::abci::global_cfg::CFG, lazy_static::lazy_static,
     ledger::staking::td_addr_to_bytes, ruc::*, serde::Deserialize, std::fs,
 };
